requirements.txt
aiogram
python-dotenv
yt-dlp
requests
beautifulsoup4 
lyricsgenius
emoji